Use Ni surface thermo library in relevant example input file.
When generating a model on Ni, we should preferentially use the Ni thermo
data when it exists. Fall back to the Pt because there's not much Ni.

Soon we will be able to correctly apply LSRs to scale from one to the other,
and this may all change. But not yet.